refactor: extract client utils in its own module
