+++ plugins/module_utils/hcloud.py
@@ -8,63 +8,13 @@ from __future__ import annotations
 import traceback
 from typing import Any
 
-from ansible.module_utils.basic import (
-    AnsibleModule as AnsibleModuleBase,
-    env_fallback,
-    missing_required_lib,
-)
+from ansible.module_utils.basic import AnsibleModule as AnsibleModuleBase, env_fallback
 from ansible.module_utils.common.text.converters import to_native
 
-from ..module_utils.vendor.hcloud import APIException, Client, HCloudException
-from ..module_utils.vendor.hcloud.actions import ActionException
+from .client import ClientException, client_check_required_lib, client_get_by_name_or_id
+from .vendor.hcloud import APIException, Client, HCloudException
+from .vendor.hcloud.actions import ActionException
 from .version import version
-
-HAS_REQUESTS = True
-HAS_DATEUTIL = True
-
-try:
-    import requests  # pylint: disable=unused-import
-except ImportError:
-    HAS_REQUESTS = False
-
-try:
-    import dateutil  # pylint: disable=unused-import
-except ImportError:
-    HAS_DATEUTIL = False
-
-
-class ClientException(Exception):
-    """An error related to the client occurred."""
-
-
-def client_check_required_lib():
-    if not HAS_REQUESTS:
-        raise ClientException(missing_required_lib("requests"))
-    if not HAS_DATEUTIL:
-        raise ClientException(missing_required_lib("python-dateutil"))
-
-
-def client_get_by_name_or_id(client: Client, resource: str, param: str | int):
-    """
-    Get a resource by name, and if not found by its ID.
-
-    :param client: Client to use to make the call
-    :param resource: Name of the resource client that implements both `get_by_name` and `get_by_id` methods
-    :param param: Name or ID of the resource to query
-    """
-    resource_client = getattr(client, resource)
-
-    result = resource_client.get_by_name(param)
-    if result is not None:
-        return result
-
-    # If the param is not a valid ID, prevent an unnecessary call to the API.
-    try:
-        int(param)
-    except ValueError as exception:
-        raise ClientException(f"resource ({resource.rstrip('s')}) does not exist: {param}") from exception
-
-    return resource_client.get_by_id(param)
 
 
 # Provide typing definitions to the AnsibleModule class
